Resolved feedback
