+++ frontend/src/modules/_shared/hooks/useIsMountedRef.ts
@@ -1,0 +1,12 @@
+import React from "react";
+
+export function useIsMountedRef() {
+    const isMountedRef = React.useRef(true);
+    React.useEffect(() => {
+        isMountedRef.current = true;
+        return () => {
+            isMountedRef.current = false;
+        };
+    }, []);
+    return isMountedRef;
+}
